Create Smallest Positive Integer
Smallest Positive Integer that can not be represented as Sum

Join the most popular course on DSA. Master Skills & Become Employable by enrolling today! 
Given an array of size N, find the smallest positive integer value that is either not presented in the array or cannot be represented as a sum(coz sum means you are adding two or more elements) of some elements from the array.
Example 1:

Input: 
N = 6
array[] = {1, 10, 3, 11, 6, 15}
Output: 
2
Explanation:
2 is the smallest integer value that cannot 
be represented as sum of elements from the array.